backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
uvicorn[standard]==0.32.0
pydantic==2.10.1
pydantic-settings==2.7.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.20
alembic==1.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.0
sqlalchemy==2.0.36